% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.assign_domain_workflow}
\alias{.assign_domain_workflow}
\title{Assign a high-level imaging workflow request to the server}
\usage{
.assign_domain_workflow(conns, method, request, symbol = NULL)
}
\description{
Assign a high-level imaging workflow request to the server
}
\keyword{internal}
