% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ds_encode}
\alias{.ds_encode}
\title{Encode R objects as URL-safe B64 JSON for DataSHIELD transport}
\usage{
.ds_encode(x)
}
\description{
Encode R objects as URL-safe B64 JSON for DataSHIELD transport
}
\keyword{internal}
