% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.generate_job_id}
\alias{.generate_job_id}
\title{Generate a client-side job id for domain-mediated dsHPC submissions}
\usage{
.generate_job_id()
}
\description{
Generate a client-side job id for domain-mediated dsHPC submissions
}
\keyword{internal}
