% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.generate_symbol}
\alias{.generate_symbol}
\title{Generate a server-side symbol for domain workflow handles}
\usage{
.generate_symbol(prefix = "dsIjob")
}
\description{
Generate a server-side symbol for domain workflow handles
}
\keyword{internal}
