% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.imaging_workflow_rollback}
\alias{.imaging_workflow_rollback}
\title{Exact rollback for a partially assigned workflow}
\usage{
.imaging_workflow_rollback(conns, symbol)
}
\description{
Exact rollback for a partially assigned workflow
}
\keyword{internal}
