% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{ds.imaging.feature_view.destroy}
\alias{ds.imaging.feature_view.destroy}
\title{Destroy an opaque imaging feature view}
\usage{
ds.imaging.feature_view.destroy(conns, symbol = "imaging_features")
}
\arguments{
\item{conns}{DSI connections object.}

\item{symbol}{Character; assigned feature-view symbol.}
}
\value{
Invisibly TRUE.
}
\description{
Destroy an opaque imaging feature view
}
