% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_assets.R, R/analysis_compat.R
\name{ds.imaging.jobs}
\alias{ds.imaging.jobs}
\alias{ds.radiomics.list}
\title{List radiomics jobs}
\usage{
ds.imaging.jobs(conns)

ds.radiomics.list(conns)
}
\arguments{
\item{conns}{DSI connections object.}
}
\value{
A dsjobs_result.
}
\description{
List radiomics jobs
}
