% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_profiles.R
\name{ds.imaging.radiomics.profile.force2d}
\alias{ds.imaging.radiomics.profile.force2d}
\title{Force-2D profile (for 2D slices)}
\usage{
ds.imaging.radiomics.profile.force2d(bin_width = 25)
}
\arguments{
\item{bin_width}{Numeric; histogram bin width (default 25).}
}
\value{
A radiomics profile spec.
}
\description{
Force-2D profile (for 2D slices)
}
