% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_compat.R, R/analysis_segmenters.R
\name{ds.segmenter.existing_mask}
\alias{ds.segmenter.existing_mask}
\alias{ds.imaging.segmenter.existing_mask}
\title{Use existing mask asset (no segmentation needed)}
\usage{
ds.segmenter.existing_mask(mask_asset)

ds.imaging.segmenter.existing_mask(mask_asset)
}
\arguments{
\item{mask_asset}{Character; asset_id or alias of existing masks.}
}
\value{
A segmenter spec.
}
\description{
Use existing mask asset (no segmentation needed)
}
