% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{ds.imaging.assets}
\alias{ds.imaging.assets}
\title{List Dataset Assets}
\usage{
ds.imaging.assets(conns, handle = "img")
}
\arguments{
\item{conns}{DSI connections object.}

\item{handle}{Character; symbol name of the imaging handle
(default \code{"img"}).}
}
\value{
Named list of per-server data.frames.
}
\description{
Calls \code{imagingAssetsDS()} on each server to retrieve the names
and types of assets available in the imaging dataset.
}
