% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_assets.R
\name{ds.imaging.capabilities}
\alias{ds.imaging.capabilities}
\title{Get radiomics capabilities from server}
\usage{
ds.imaging.capabilities(conns)
}
\arguments{
\item{conns}{DSI connections object.}
}
\value{
Named list of per-server capabilities.
}
\description{
Get radiomics capabilities from server
}
