% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{ds.imaging.datasets}
\alias{ds.imaging.datasets}
\title{List Available Imaging Datasets}
\usage{
ds.imaging.datasets(conns)
}
\arguments{
\item{conns}{DSI connections object.}
}
\value{
Named list of per-server data.frames with available datasets.
}
\description{
Calls \code{imagingListDatasetsDS()} on each server to retrieve the
list of available imaging datasets.
}
