% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_assets.R
\name{ds.imaging.jobs}
\alias{ds.imaging.jobs}
\title{List radiomics jobs}
\usage{
ds.imaging.jobs(conns)
}
\arguments{
\item{conns}{DSI connections object.}
}
\value{
A dshpc_result; printed as a formatted job table.
}
\description{
List radiomics jobs
}
\examples{
\donttest{
# conns <- DSI::datashield.login(...)  # live DataSHIELD session
ds.imaging.jobs(conns)
}
}
