% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_admin.R, R/analysis_compat.R
\name{ds.imaging.models}
\alias{ds.imaging.models}
\alias{ds.radiomics.models}
\title{List installed models on the server}
\usage{
ds.imaging.models(conns)

ds.radiomics.models(conns)
}
\arguments{
\item{conns}{DSI connections object.}
}
\value{
Named list with per-server model listings.
}
\description{
List installed models on the server
}
