% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_workflows_collection.R
\name{ds.imaging.radiomics.collection_publish}
\alias{ds.imaging.radiomics.collection_publish}
\title{Publish a completed collection workflow}
\usage{
ds.imaging.radiomics.collection_publish(conns, symbol)
}
\arguments{
\item{conns}{DSI connections object.}

\item{symbol}{Character; server-side collection workflow symbol.}
}
\value{
A list containing \code{state} and \code{asset_id}; or a named list
of these responses for multiple servers.
}
\description{
Publish a completed collection workflow
}
