% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_compat.R, R/analysis_profiles.R
\name{ds.radiomics.profile.ibsi_ct_3d}
\alias{ds.radiomics.profile.ibsi_ct_3d}
\alias{ds.imaging.radiomics.profile.ibsi_ct_3d}
\title{IBSI-compliant CT 3D profile}
\usage{
ds.radiomics.profile.ibsi_ct_3d(bin_width = 25)

ds.imaging.radiomics.profile.ibsi_ct_3d(bin_width = 25)
}
\arguments{
\item{bin_width}{Numeric; histogram bin width (default 25).}
}
\value{
A radiomics profile spec.
}
\description{
IBSI-compliant CT 3D profile
}
