% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_segmenters.R
\name{ds.imaging.segmenter.monai_bundle}
\alias{ds.imaging.segmenter.monai_bundle}
\title{MONAI bundle segmenter}
\usage{
ds.imaging.segmenter.monai_bundle(bundle_name)
}
\arguments{
\item{bundle_name}{Character; registered bundle name.}
}
\value{
A segmenter spec.
}
\description{
MONAI bundle segmenter
}
