% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_segmenters.R
\name{ds.imaging.segmenter.nnunet}
\alias{ds.imaging.segmenter.nnunet}
\title{nnU-Net v2 segmenter}
\usage{
ds.imaging.segmenter.nnunet(model_name, fold = "all")
}
\arguments{
\item{model_name}{Character; registered model name.}

\item{fold}{Character; fold to use (default "all").}
}
\value{
A segmenter spec.
}
\description{
nnU-Net v2 segmenter
}
